[AT-1788] - Move ds sim tests from ska-mid-itf repo.
This change migrates the ds sim tests from the ska-mid-itf repo to here.
This enables us to run the tests prior to making new releases and also simplifies the test execution in the ska-mid-itf repo considerably.
The previous tests in the repo were not being executed in CI/CD and their functionality was duplicated in the new tests so they have been removed.
The usage of helm test was also removed from the k8s-test-runner stage to make room for the new tests. The helm tests are no longer neccesary in this stage since they only ping the HTTP endpoint of the dish structure simulator and this is already done in the python tests.

1. Made all deployments use the loadbalancer servicetype since this is supported by ska-cicd-deploy-minikube.
2. Added python-lint to pipeline.